fix(frontend): mysql授权单据详情页交互调整 #4037 # Reviewed, transaction id: 6163
