+++ dbm-ui/frontend/src/locales/zh-cn.json
@@ -2871,5 +2871,4 @@
     "继续执行单据后无法撤回，请谨慎操作！": "继续执行单据后无法撤回，请谨慎操作！",
     "终止单据后无法撤回，请谨慎操作！": "终止单据后无法撤回，请谨慎操作！",
     "这行勿动！新增翻译请在上一行添加！": ""
-
 }
