+++ dbm-ui/frontend/src/views/tickets/common/components/demand-factory/Index.vue
@@ -37,6 +37,7 @@
   import RiakReboot from './bigdata/RiakReboot.vue';
   import InfluxdbOperations from './influxdb/Operations.vue';
   import InfluxdbReplace from './influxdb/Replace.vue';
+  import MySQLAuthorizeRule from './mysql/authorize-rule/Index.vue';
   import MySQLChecksum from './mysql/Checksum.vue';
   import MySQLClone from './mysql/Clone.vue';
   import MySQLClusterOperation from './mysql/ClusterOperation.vue';
@@ -52,7 +53,6 @@
   import MySQLMasterSlaveSwitch from './mysql/MasterSlaveSwitch.vue';
   import MySQLMigrateCluster from './mysql/MigrateCluster.vue';
   import MysqlOpenArea from './mysql/OpenArea.vue';
-  import MySQLOperation from './mysql/Operation.vue';
   import MySQLProxyAdd from './mysql/ProxyAdd.vue';
   import MySQLProxySwitch from './mysql/ProxySwitch.vue';
   import MySQLRename from './mysql/Rename.vue';
@@ -306,7 +306,7 @@
       return RedisOperation;
     }
     if (mysqlTicketType.includes(ticketType)) {
-      return MySQLOperation;
+      return MySQLAuthorizeRule;
     }
     // influxdb 禁用/重启/启用/删除
     if (influxdbTypes.includes(ticketType)) {
